refactor: components 폴더에 있는 불명확한 폴더 구조 개선
