+++ src/components/Header/Logo/logo.css
@@ -1,0 +1,5 @@
+.logo {
+  width: 100%;
+  padding: 30px 0;
+  text-align: center;
+}
