+++ src/components/Header/UtilMenu/utilMenu.css
@@ -1,0 +1,43 @@
+.util-menu {
+  position: absolute;
+  top: 35px;
+  right: -65px;
+
+  display: flex;
+  align-items: center;
+
+  height: 35px;
+
+  font-size: 13px;
+  color: #666;
+}
+
+.util-menu > .sign {
+  position: relative;
+}
+
+.util-menu > .sign ~ .sign {
+  margin-left: 30px;
+}
+
+.util-menu > .sign ~ .sign::before {
+  content: "";
+
+  position: absolute;
+  top: 50%;
+  right: 125%;
+  transform: translate(-50%, -50%);
+
+  display: block;
+
+  width: 3px;
+  height: 3px;
+  margin-top: -2px;
+
+  background-color: #ccc;
+  border-radius: 50%;
+}
+
+.util-menu > .global-subway-icon {
+  margin-left: 15px;
+}
