+++ src/pages/Home/home.css
@@ -1,3 +1,53 @@
+/* global style */
+a:link {
+  text-decoration: none;
+}
+
+a:visited,
+a:active {
+  color: inherit;
+}
+
+/* Header 전체 Style */
+.header {
+  position: absolute;
+  top: 0;
+  left: 0;
+
+  width: 100%;
+  height: 165px;
+
+  background-color: #fff;
+
+  transition: height 0.3s;
+}
+
+.header-over {
+  height: 405px;
+  box-shadow: 0 3px 5px rgb(0 0 0 / 10%);
+}
+
+.header::after {
+  content: "";
+
+  position: absolute;
+  top: 164px;
+  left: 0;
+
+  width: 100%;
+  height: 1px;
+
+  background-color: #e5e5e5;
+}
+
+.header-inner {
+  position: relative;
+  max-width: 65%;
+  height: 100%;
+  margin: 0 auto;
+}
+
+/* Template */
 .main {
   width: 100%;
   height: 350px;
